requirements.txt
EasyDict>=1.7
opencv-python>=3.4.1.15
shapely>=1.6.4
Cython
scipy
pandas
pyyaml
json_tricks
scikit-image
tensorflow
tensorboard
tensorboardX
pymvg
numexpr
h5py
tqdm
torchvision
GitPython
